remove todos
